+++ myokit/formats/ansic/_ewriter.py
@@ -21,20 +21,47 @@ class AnsiCExpressionWriter(PythonExpressionWriter):
 
     def set_condition_function(self, func=None):
         """
-        Sets a function name to use for if statements
+        Sets a function name to use for :class:`myokit.If`; if not set the
+        ternary operatur will be used.
 
-        By setting func to None you can revert back to the default behavior
-         (the ternary operator). Any other value will be interpreted as the
-         name of a C function taking arguments (condition, value_if_true,
-         value_if_false).
+        If given, the function arguments should be ``(condition, value_if_true,
+         value_if_false)``. To revert to using the ternary operator, call with
+         ``func=None``.
         """
         self._fcond = func
 
+    def _ex_prefix(self, e, op):
+        # PrefixPlus and PrefixMinus. No simplifications should be made here
+        # for PrefixPlus, see https://github.com/myokit/myokit/issues/1054
+        x = self.ex(e[0])
+        return f'{op}({x})' if e.bracket(e[0]) or x[0] == op else f'{op}{x}'
+
+    def _ex_infix_comparison(self, e, op):
+        # For equals etc
+        return f'({self.ex(e[0])} {op} {self.ex(e[1])})'
+
+    def _ex_infix_logical(self, e, op):
+        # For and and or
+        return f'({self.ex(e[0])} {op} {self.ex(e[1])})'
+
     #def _ex_name(self, e):
     #def _ex_derivative(self, e):
+
+    def _ex_initial_value(self, e):
+        # These are disabled by default, but enabled here to support CVODES
+        # sensitivity calculations.
+        return self._flhs(e)
+
+    def _ex_partial_derivative(self, e):
+        # These are disabled by default, but enabled here to support CVODES
+        # sensitivity calculations.
+        return self._flhs(e)
+
     #def _ex_number(self, e):
+
     #def _ex_prefix_plus(self, e):
     #def _ex_prefix_minus(self, e):
+
     #def _ex_plus(self, e):
     #def _ex_minus(self, e):
     #def _ex_multiply(self, e):
@@ -43,17 +70,20 @@ class AnsiCExpressionWriter(PythonExpressionWriter):
     def _ex_quotient(self, e):
         # Note that this _must_ round towards minus infinity!
         # See myokit.Quotient !
+        # No extra brackets needed: Function
         return self.ex(myokit.Floor(myokit.Divide(e[0], e[1])))
 
     def _ex_remainder(self, e):
         # Note that this _must_ use the same round-to-neg-inf convention as
         # myokit.Quotient! Implementation below is consistent with Python
-        # convention:
-        return self.ex(myokit.Minus(
-            e[0], myokit.Multiply(e[1], myokit.Quotient(e[0], e[1]))))
+        # convention.
+        # Extra brackets needed! Minus has lower precedence than division.
+        i = myokit.Minus(e[0], myokit.Multiply(e[1],
+                myokit.Floor(myokit.Divide(e[0], e[1]))))
+        return f'({self.ex(i)})'
 
     def _ex_power(self, e):
-        return 'pow(' + self.ex(e[0]) + ', ' + self.ex(e[1]) + ')'
+        return f'pow({self.ex(e[0])}, {self.ex(e[1])})'
 
     #def _ex_sqrt(self, e):
     #def _ex_sin(self, e):
@@ -67,7 +97,9 @@ class AnsiCExpressionWriter(PythonExpressionWriter):
     def _ex_log(self, e):
         if len(e) == 1:
             return self._ex_function(e, 'log')
-        return '(log(' + self.ex(e[0]) + ') / log(' + self.ex(e[1]) + '))'
+        # Always add brackets: Parent was expecting a function so will never
+        # have added them.
+        return f'(log({self.ex(e[0])}) / log({self.ex(e[1])}))'
 
     #def _ex_log10(self, e):
     #def _ex_floor(self, e):
@@ -75,9 +107,6 @@ class AnsiCExpressionWriter(PythonExpressionWriter):
 
     def _ex_abs(self, e):
         return self._ex_function(e, 'fabs')
-
-    def _ex_not(self, e):
-        return '!(' + self.ex(e[0]) + ')'
 
     #def _ex_equal(self, e):
     #def _ex_not_equal(self, e):
@@ -87,32 +116,47 @@ class AnsiCExpressionWriter(PythonExpressionWriter):
     #def _ex_less_equal(self, e):
 
     def _ex_and(self, e):
-        return self._ex_infix_condition(e, '&&')
+        return self._ex_infix_logical(e, '&&')
 
     def _ex_or(self, e):
-        return self._ex_infix_condition(e, '||')
+        return self._ex_infix_logical(e, '||')
+
+    def _ex_not(self, e):
+        # C conditions all have brackets, so don't add more
+        if isinstance(e[0], (myokit.Condition)):
+            return f'(!{self.ex(e[0])})'
+        # But do add more if the user's being silly
+        return f'(!({self.ex(e[0])}))'
 
     def _ex_if(self, e):
-        ite = (self.ex(e._i), self.ex(e._t), self.ex(e._e))
-        if self._fcond is None:
-            return '(%s ? %s : %s)' % ite
-        else:
-            return '%s(%s, %s, %s)' % ((self._fcond,) + ite)
+        _if, _then, _else = self.ex(e._i), self.ex(e._t), self.ex(e._e)
+        # If i is not a condtion (which always gets brackets from this writer)
+        # then add brackets
+        if not isinstance(e._i, myokit.Condition):
+            _if = f'({_if})'
+
+        # Use if-then-else function?
+        if self._fcond is not None:
+            return f'{self._fcond}({_if}, {_then}, {_else})'
+
+        # Default: use ternary operator
+        return f'({_if} ? {_then} : {_else})'
 
     def _ex_piecewise(self, e):
+        # Render ifs; add extra bracket if not a condition (see _ex_if)
+        _ifs = [self.ex(x) if isinstance(x, myokit.Condition)
+                else f'({self.ex(x)})' for x in e._i]
+        _thens = [self.ex(x) for x in e._e]
+
         s = []
-        n = len(e._i)
-        if self._fcond is None:
-            for i in range(0, n):
-                s.append('(%s ? %s : ' % (self.ex(e._i[i]), self.ex(e._e[i])))
-            s.append(self.ex(e._e[n]))
-            s.append(')' * n)
+        n = len(_ifs)
+        if self._fcond is not None:
+            for _if, _then in zip(_ifs, _thens):
+                s.append(f'{self._fcond}({_if}, {_then}, ')
         else:
-            for i in range(0, n):
-                s.append(
-                    '%s(%s, %s, ' % (
-                        self._fcond, self.ex(e._i[i]), self.ex(e._e[i])))
-            s.append(self.ex(e._e[n]))
-            s.append(')' * n)
+            for _if, _then in zip(_ifs, _thens):
+                s.append(f'({_if} ? {_then} : ')
+        s.append(_thens[-1])
+        s.append(')' * len(_ifs))
         return ''.join(s)