+++ myokit/formats/cpp/_ewriter.py
@@ -12,4 +12,10 @@ class CppExpressionWriter(AnsiCExpressionWriter):
     This :class:`ExpressionWriter <myokit.formats.ExpressionWriter>` translates
     myokit :class:`expressions <myokit.Expression>` to their C++ equivalent.
     """
-    pass
+    def _ex_initial_value(self, e):
+        raise NotImplementedError(
+            'Initial values are not supported by this expression writer.')
+
+    def _ex_partial_derivative(self, e):
+        raise NotImplementedError(
+            'Partial derivatives are not supported by this expression writer.')