+++ myokit/tests/test_formats_latex.py
@@ -1,0 +1,164 @@
+class LatexExpressionWriterTest(unittest.TestCase):
+    """ Test the Latex ewriter class. """
+
+    def test_all(self):
+        w = myokit.formats.latex.LatexExpressionWriter()
+
+        model = myokit.Model()
+        component = model.add_component('c')
+        avar = component.add_variable('a')
+
+        # Model needs to be validated --> sets unames
+        avar.set_rhs(12)
+        avar.set_binding('time')
+        model.validate()
+
+        # Name
+        a = myokit.Name(avar)
+        self.assertEqual(w.ex(a), '\\text{a}')
+        # Number with unit
+        b = myokit.Number('12', 'pF')
+        self.assertEqual(w.ex(b), '12.0')
+
+        # Prefix plus
+        x = myokit.PrefixPlus(b)
+        self.assertEqual(w.ex(x), '12.0')
+        # Prefix minus
+        x = myokit.PrefixMinus(b)
+        self.assertEqual(w.ex(x), '\\left(-12.0\\right)')
+        # Prefix minus with bracket
+        x = myokit.PrefixMinus(myokit.Plus(a, b))
+        self.assertEqual(
+            w.ex(x), '\\left(-\\left(\\text{a}+12.0\\right)\\right)')
+
+        # Plus
+        x = myokit.Plus(a, b)
+        self.assertEqual(w.ex(x), '\\text{a}+12.0')
+        # Minus
+        x = myokit.Minus(a, b)
+        self.assertEqual(w.ex(x), '\\text{a}-12.0')
+        # Multiply
+        x = myokit.Multiply(a, b)
+        self.assertEqual(w.ex(x), '\\text{a}*12.0')
+        # Divide
+        x = myokit.Divide(a, b)
+        self.assertEqual(w.ex(x), '\\frac{\\text{a}}{12.0}')
+
+        # Quotient
+        # Not supported in latex!
+        x = myokit.Quotient(a, b)
+        self.assertEqual(
+            w.ex(x), '\\left\\lfloor\\frac{\\text{a}}{12.0}\\right\\rfloor')
+        # Remainder
+        x = myokit.Remainder(a, b)
+        self.assertEqual(w.ex(x), '\\bmod\\left(\\text{a},12.0\\right)')
+
+        # Power
+        x = myokit.Power(a, b)
+        self.assertEqual(w.ex(x), '\\text{a}^{12.0}')
+        # Power with brackets
+        x = myokit.Power(myokit.Plus(a, b), b)
+        self.assertEqual(w.ex(x), '\\left(\\text{a}+12.0\\right)^{12.0}')
+        # Sqrt
+        x = myokit.Sqrt(b)
+        self.assertEqual(w.ex(x), '\\sqrt{12.0}')
+        # Exp
+        x = myokit.Exp(a)
+        self.assertEqual(w.ex(x), '\\exp\\left(\\text{a}\\right)')
+        # Log(a)
+        x = myokit.Log(b)
+        self.assertEqual(w.ex(x), '\\log\\left(12.0\\right)')
+        # Log(a, b)
+        x = myokit.Log(a, b)
+        self.assertEqual(w.ex(x), '\\log_{12.0}\\left(\\text{a}\\right)')
+        # Log10
+        x = myokit.Log10(b)
+        self.assertEqual(w.ex(x), '\\log_{10.0}\\left(12.0\\right)')
+
+        # Sin
+        x = myokit.Sin(b)
+        self.assertEqual(w.ex(x), '\\sin\\left(12.0\\right)')
+        # Cos
+        x = myokit.Cos(b)
+        self.assertEqual(w.ex(x), '\\cos\\left(12.0\\right)')
+        # Tan
+        x = myokit.Tan(b)
+        self.assertEqual(w.ex(x), '\\tan\\left(12.0\\right)')
+        # ASin
+        x = myokit.ASin(b)
+        self.assertEqual(w.ex(x), '\\arcsin\\left(12.0\\right)')
+        # ACos
+        x = myokit.ACos(b)
+        self.assertEqual(w.ex(x), '\\arccos\\left(12.0\\right)')
+        # ATan
+        x = myokit.ATan(b)
+        self.assertEqual(w.ex(x), '\\arctan\\left(12.0\\right)')
+
+        # Floor
+        x = myokit.Floor(b)
+        self.assertEqual(w.ex(x), '\\left\\lfloor{12.0}\\right\\rfloor')
+        # Ceil
+        x = myokit.Ceil(b)
+        self.assertEqual(w.ex(x), '\\left\\lceil{12.0}\\right\\rceil')
+        # Abs
+        x = myokit.Abs(b)
+        self.assertEqual(w.ex(x), '\\lvert{12.0}\\rvert')
+
+        # Equal
+        x = myokit.Equal(a, b)
+        self.assertEqual(w.ex(x), '\\left(\\text{a}=12.0\\right)')
+        # NotEqual
+        x = myokit.NotEqual(a, b)
+        self.assertEqual(w.ex(x), '\\left(\\text{a}\\neq12.0\\right)')
+        # More
+        x = myokit.More(a, b)
+        self.assertEqual(w.ex(x), '\\left(\\text{a}>12.0\\right)')
+        # Less
+        x = myokit.Less(a, b)
+        self.assertEqual(w.ex(x), '\\left(\\text{a}<12.0\\right)')
+        # MoreEqual
+        x = myokit.MoreEqual(a, b)
+        self.assertEqual(w.ex(x), '\\left(\\text{a}\\geq12.0\\right)')
+        # LessEqual
+        x = myokit.LessEqual(a, b)
+        self.assertEqual(w.ex(x), '\\left(\\text{a}\\leq12.0\\right)')
+
+        # Not
+        cond1 = myokit.parse_expression('5 > 3')
+        cond2 = myokit.parse_expression('2 < 1')
+        x = myokit.Not(cond1)
+        self.assertEqual(
+            w.ex(x), '\\not\\left(\\left(5.0>3.0\\right)\\right)')
+        # And
+        x = myokit.And(cond1, cond2)
+        self.assertEqual(
+            w.ex(x),
+            '\\left(\\left(5.0>3.0\\right)\\and'
+            '\\left(2.0<1.0\\right)\\right)')
+        # Or
+        x = myokit.Or(cond1, cond2)
+        self.assertEqual(
+            w.ex(x),
+            '\\left(\\left(5.0>3.0\\right)\\or'
+            '\\left(2.0<1.0\\right)\\right)')
+        # If
+        x = myokit.If(cond1, a, b)
+        self.assertEqual(
+            w.ex(x), 'if\\left(\\left(5.0>3.0\\right),\\text{a},12.0\\right)')
+        # Piecewise
+        c = myokit.Number(1)
+        x = myokit.Piecewise(cond1, a, cond2, b, c)
+        self.assertEqual(
+            w.ex(x),
+            'piecewise\\left(\\left(5.0>3.0\\right),\\text{a},'
+            '\\left(2.0<1.0\\right),12.0,1.0\\right)')
+
+        # Test fetching using ewriter method
+        w = myokit.formats.ewriter('latex')
+        self.assertIsInstance(w, myokit.formats.latex.LatexExpressionWriter)
+
+        # Test without a Myokit expression
+        self.assertRaisesRegex(
+            ValueError, 'Unknown expression type', w.ex, 7)
+
+